Move search components one level up
