+++ client/src/components/Connections/components/Search.js
@@ -4,11 +4,11 @@ import {withRouter} from 'react-router-dom'
 import {withHandlers, withState} from 'recompose'
 import {compose} from 'redux'
 import {connect} from 'react-redux'
-import {updateValue} from '../../../../actions/sharedActions'
+import {updateValue} from '../../../actions/sharedActions'
 import {Form, FormGroup, Label, Input, Button} from 'reactstrap'
 import type {ContextRouter} from 'react-router'
 
-import EntitySearchWrapper from '../../dataWrappers/EntitySearchWrapper'
+import EntitySearchWrapper from '../dataWrappers/EntitySearchWrapper'
 import './Search.css'
 
 const checkEnter = (callback) => {
